Fix CMake package include path. Fix all a conversion from 'double' to 'float' warnings in code. Fix check robot and motor speed in negative field. Increase error (invalid argument) message buffer size to 100 chars. Fix number of motor in error message (increase by 1). Fix arguments type of set_motor_pid() and set_motors_pid() functions from float to unsigned int. Change numerical constant literals from double to int or float. Change speed limits.

+++ tests/main.cpp
@@ -19,7 +19,7 @@ int main( int argc, char **argv )
     while (timer.msecsElapsed() <= test_duration)
     {
         try {
-            robotino.set_robot_speed(0.1, 0, 0);
+            robotino.set_robot_speed((float)0.1, (float)0, (float)0);
         }
         catch(const std::invalid_argument& e) {
             std::cout << "set_robot_speed() failed with error: " << e.what();
